+++ packages/skills/src/application/useCases/findSkillBySlug/FindSkillBySlugUsecase.spec.ts
@@ -14,7 +14,7 @@ import {
 import { v4 as uuidv4 } from 'uuid';
 import { skillFactory } from '../../../../test/skillFactory';
 import { SkillService } from '../../services/SkillService';
-import { FindSkillBySlugUsecase } from './findSkillBySlug.usecase';
+import { FindSkillBySlugUsecase } from './FindSkillBySlugUsecase';
 
 describe('FindSkillBySlugUsecase', () => {
   let usecase: FindSkillBySlugUsecase;
